api-gateway/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
httpx==0.25.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0